Moved the meta-proxy to package "ipvs"
